backend_functions/py/ascii/requirements.txt
Pillow
wget
hashlib
json
argparse
opencv-python
numpy